rename legalelo package to core
