+++ .gitignore
@@ -8,3 +8,4 @@ causaganha/data/json_test_output/
 causaganha/data/json_cli_test/
 causaganha/data/json_extract_test/
 causaganha/data/json_test_extractor/
+site/
