--- causaganha/legalelo/__init__.py
@@ -1,1 +1,0 @@
-# This file makes 'legalelo' a sub-package of 'causaganha'.
