+++ causaganha/tests/test_utils.py
@@ -9,7 +9,7 @@ PROJECT_ROOT = Path(__file__).resolve().parent.parent.parent
 if str(PROJECT_ROOT) not in sys.path:
     sys.path.insert(0, str(PROJECT_ROOT))
 
-from causaganha.legalelo.utils import normalize_lawyer_name, validate_decision
+from causaganha.core.utils import normalize_lawyer_name, validate_decision
 
 # Suppress logging output during tests unless specifically testing for it
 logging.disable(logging.CRITICAL)
@@ -69,91 +69,91 @@ class TestValidateDecision(unittest.TestCase):
     def test_missing_numero_processo(self):
         invalid = self.valid_decision.copy()
         del invalid["numero_processo"]
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'numero_processo' is missing or empty.")
 
     def test_bad_numero_processo_format(self):
         invalid = self.valid_decision.copy()
         invalid["numero_processo"] = "123.456" # Too short / wrong format
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with(f"Validation failed: 'numero_processo' ({invalid['numero_processo']}) does not match pattern [\\d.-]{{15,25}}.")
 
     def test_numero_processo_not_string(self):
         invalid = self.valid_decision.copy()
         invalid["numero_processo"] = 12345
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with(f"Validation failed: 'numero_processo' is not a string (got {type(invalid['numero_processo'])}). Value: {invalid['numero_processo']}")
 
     def test_missing_partes(self):
         invalid = self.valid_decision.copy()
         del invalid["partes"]
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with(f"Validation failed: 'partes' is missing or not a dictionary (got <class 'NoneType'>).")
 
     def test_partes_not_dict(self):
         invalid = self.valid_decision.copy()
         invalid["partes"] = "not a dict"
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with(f"Validation failed: 'partes' is missing or not a dictionary (got {type(invalid['partes'])}).")
 
     def test_missing_requerente(self):
         invalid = self.valid_decision.copy()
         invalid["partes"] = {"requerido": ["Some Recipient"]}
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'partes.requerente' is missing or empty.")
 
     def test_empty_requerente_list(self):
         invalid = self.valid_decision.copy()
         invalid["partes"]["requerente"] = []
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'partes.requerente' is missing or empty.")
 
     def test_empty_requerente_string(self): # Assuming string is allowed based on current validate_decision
         invalid = self.valid_decision.copy()
         invalid["partes"]["requerente"] = ""
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'partes.requerente' is missing or empty.")
 
     def test_requerente_wrong_type(self):
         invalid = self.valid_decision.copy()
         invalid["partes"]["requerente"] = 123
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with(f"Validation failed: 'partes.requerente' is not a list or string (got {type(invalid['partes']['requerente'])}).")
 
     def test_missing_requerido(self): # Similar tests for 'requerido'
         invalid = self.valid_decision.copy()
         invalid["partes"] = {"requerente": ["Some Applicant"]}
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'partes.requerido' is missing or empty.")
 
     def test_missing_resultado(self):
         invalid = self.valid_decision.copy()
         del invalid["resultado"]
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'resultado' is missing or empty.")
 
     def test_empty_resultado_string(self):
         invalid = self.valid_decision.copy()
         invalid["resultado"] = ""
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             mock_log.assert_called_with("Validation failed: 'resultado' is missing or empty.")
 
     def test_resultado_wrong_type(self):
         invalid = self.valid_decision.copy()
         invalid["resultado"] = [] # Empty list makes 'not resultado' True
-        with patch.object(logging.getLogger('causaganha.legalelo.utils'), 'warning') as mock_log:
+        with patch.object(logging.getLogger('causaganha.core.utils'), 'warning') as mock_log:
             self.assertFalse(validate_decision(invalid))
             # The first check to fail for an empty list is 'missing or empty'
             mock_log.assert_called_with("Validation failed: 'resultado' is missing or empty.")
