+++ docs/faq.md
@@ -1,0 +1,10 @@
+# FAQ e Limitações
+
+**Como configurar a chave da API do Gemini?**
+
+Defina o segredo `GEMINI_API_KEY` nas configurações do repositório.
+
+**Quais são as principais limitações?**
+
+- Dependência de cota da API do Gemini.
+- Apenas o TJRO está implementado.
