+++ docs/index.md
@@ -1,0 +1,6 @@
+# CausaGanha Documentation
+
+Bem-vindo à documentação oficial do CausaGanha.
+
+- [Tutorial Rápido](quickstart.md)
+- [FAQ e Limitações](faq.md)
