+++ docs/quickstart.md
@@ -1,0 +1,8 @@
+# Tutorial Rápido
+
+Este tutorial demonstra como rodar o pipeline completo em minutos.
+
+```bash
+pip install -r requirements.txt
+python -m causaganha.core.pipeline run --date 2025-06-01
+```
