+++ requirements.txt
@@ -11,3 +11,4 @@ pre-commit==3.7.1
 ruff==0.4.7
 google-api-python-client==2.132.0
 google-auth==2.29.0
+mkdocs==1.5.3
